+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - `Innmind\TimeWarp\Halt` is now a final class
+- Requires PHP `8.4`
 
 ## 4.1.1 - 2025-08-18